% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_classweights.R
\name{get_classweights}
\alias{get_classweights}
\title{get_classweights}
\usage{
get_classweights(classes)
}
\arguments{
\item{classes}{The output of get_closepair_classes. A list of the class of each close pair.}
}
\description{
This function takes a set of close pair classes (output of get_closepair_classes)
and computes the weight of each close pair based on the size of the class.
}
\examples{
get_classweights(classes)
}
