% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_closepair_classes.R
\name{get_closepair_classes}
\alias{get_closepair_classes}
\title{get_closepairs_classes}
\usage{
get_closepair_classes(close_pairs)
}
\arguments{
\item{close_pairs}{A matrix of close pairs; the output of get_closepairs_(distance, fraction, auto)}
}
\description{
This function takes a set of close pairs and sorts them into bushes (equivalence classes defined by the relation
of being a close pair). So if Samples 1 and 2 are close pairs and Samples 2 and 3 are close pairs, then Samples 1, 2, and 3
are all in the same close pair class.
}
\examples{
get_closepair_classes(close_pairs)
}
