% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_closepairs_distance.R
\name{get_closepairs_distance}
\alias{get_closepairs_distance}
\title{get_closepairs_distance}
\usage{
get_closepairs_distance(distance_matrix, distance_cutoff, show_hist, verbose)
}
\arguments{
\item{distance_matrix}{A distance matrix}

\item{distance_cutoff}{An inclusive distance cutoff.}

\item{show_hist}{Boolean indicator for whether or not to plot a histogram of the distances and cutoff. Uses ggplot defaults for bins.}

\item{verbose}{TRUE to print distance cutoff, FALSE otherwise}
}
\description{
This function takes a distance matrix and finds all pairs of
samples that are within a specified distance cutoff. The pairs are
returned as an nx2 matrix, where n is the number of close pairs, and
the elements of the matrix are the indices of the samples in that close
pair in the distance matrix.
}
\examples{
get_closepairs_distance(distance_matrix, 0.0005)
}
