% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scores.R
\name{get_scores}
\alias{get_scores}
\title{get_scores}
\usage{
get_scores(
  pa_matrix,
  distance_matrix,
  closepair_method,
  closepair_params,
  blocksize,
  verbose
)
}
\arguments{
\item{pa_matrix}{A presence-absence matrix for the trait in question (i.e. gene, allele, phenotype, etc.). Rows are trait, columns are sample. Rownames should be trait names, colnames should be sample names.}

\item{distance_matrix}{A distance matrix. Rownames and colnames should be sample names and should be in the same order as the columns in pa_matrix.}

\item{closepair_method}{Either 'distance', 'fraction', 'fixednumber', or 'auto'.}

\item{closepair_params}{A list. For 'distance', the distance cutoff and show_hist. For 'fraction', the fraction and show_hist. For 'fixednumber', the number of close pairs, a random seed, and show_hist. For 'auto', the which_valley, nbins, maxvalleyheight, and show_hist. verbose is inherited from the main function call.}

\item{blocksize}{The number of traits to consider at once in the computation. Changing this number may help speed up the computation. Try 100 first.}

\item{verbose}{TRUE for progress reports. FALSE for no reports.}
}
\description{
This function takes a distance matrix, a presence-absence matrix, a choice of
how to determine close pairs, and a list of relevant parameters for that choice,
and returns the close pair scores.
}
\examples{

}
