% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scores_pa_closepairs.R
\name{get_scores_pa_close_pairs}
\alias{get_scores_pa_close_pairs}
\title{get_scores_pa_close_pairs}
\usage{
get_scores_pa_close_pairs(
  pa_matrix,
  close_pairs,
  classweights,
  blocksize,
  verbose
)
}
\arguments{
\item{pa_matrix}{A presence-absence matrix for the trait in question (i.e. gene, allele, phenotype, etc.). Rows are trait, columns are sample.}

\item{close_pairs}{A matrix of close pairs. Output from get_close_pairs_(distance, fraction, auto).}

\item{classweights}{A vector of weights for each close pair. Output from get_classweights.}

\item{blocksize}{The number of traits to consider at once in the computation. Changing this number may help speed up the computation. Try 100 first.}

\item{verbose}{TRUE to print progress, FALSE to not print progress.}
}
\description{
This function takes a presence-absence matrix and a set of close pairs and
computes the close-pair score for each pair of objects in the presence-absence matrix (a pair of genes, a pair of snps, etc.).
}
\examples{
get_scores_pa_close_pairs(pa_matrix, close_pairs, classweights, 100, T)
}
